strucutre for the cart using DI
