--- lubby/lib/app/src/modules/add_products/models/product.dart
@@ -1,17 +1,0 @@
-import 'dart:ffi';
-
-import 'package:mobx/mobx.dart';
-
-class Product {
-  @observable
-  String name;
-
-  @action
-  setName(String value) => name = value;
-
-  @observable
-  double price;
-
-  @action
-  setPrice(double value) => price = value;
-}
