--- lubby/lib/app/src/modules/product/product_page.dart
@@ -1,26 +1,0 @@
-import 'package:flutter_modular/flutter_modular.dart';
-import 'package:flutter/material.dart';
-import 'package:lubby/app/src/modules/product/stores/product_store.dart';
-
-class ProductsPage extends StatefulWidget {
-  final String title;
-  const ProductsPage({Key key, this.title = 'ProductPage'}) : super(key: key);
-  @override
-  ProductPageState createState() => ProductPageState();
-}
-
-class ProductPageState extends State<ProductsPage> {
-  final ProductStore store = Modular.get();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Column(
-        children: <Widget>[],
-      ),
-    );
-  }
-}
