--- lubby/lib/app/src/modules/product/stores/product_store.dart
@@ -1,15 +1,0 @@
-import 'package:mobx/mobx.dart';
-
-part 'product_store.g.dart';
-
-class ProductStore = _ProductStoreBase with _$ProductStore;
-abstract class _ProductStoreBase with Store {
-
-  @observable
-  int value = 0;
-
-  @action
-  void increment() {
-    value++;
-  } 
-}
