+++ lubby/lib/app/src/routes/routes.dart
@@ -1,4 +1,3 @@
 const String home = '/home';
 const String products = '/products';
 const String cart = '/cart';
-const String add = '/add';
