+++ lubby/test/app/src/modules/add_products/addProducts_module_test.dart
@@ -1,10 +1,8 @@
-import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_modular_test/flutter_modular_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:lubby/app/src/modules/add_products/addProducts_module.dart';
- 
-void main() {
+import 'package:lubby/app/src/modules/add_products/add_products_module.dart';
 
+void main() {
   setUpAll(() {
     initModule(AddProductsModule());
   });
