+++ lubby/test/app/src/modules/add_products/addProducts_store_test.dart
@@ -1,12 +1,10 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:lubby/app/src/modules/add_products/stores/addProducts_store.dart';
+import 'package:lubby/app/src/modules/add_products/stores/add_products_store.dart';
 
 void main() {
   AddProductsStore store;
 
-  setUpAll(() {
-    store = AddProductsStore();
-  });
+  setUpAll(() {});
 
   test('increment count', () async {});
 }
