+++ lubby/test/app/src/modules/home/product/product_module_test.dart
@@ -1,9 +1,5 @@
-import 'package:flutter_modular_test/flutter_modular_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:lubby/app/src/modules/product/product_module.dart';
 
 void main() {
-  setUpAll(() {
-    initModule(ProductModule());
-  });
+  setUpAll(() {});
 }
